Trying to fix my newest container box
Trying to fix my newest container box with proper CSS. It wasn't very good for mobile pages, as the posts went out of the box.

Fixing with an overflow property.

Also limiting the number of hits in the array to 4, but it's hardcoded atm. Will fix this later on, it's just for the CSS atm