Renamed list-making js file
Renamed the JS file for creating the list of products from an array + improved the script.

It's now able to differentiate between what page I go to based on the page title. It's also able to switch between pages if I want to load the next or previous page

The CSS which is for index.html is set inside it's own file, which means I can use the same HTML in my divs and style them differently. A lucky accident