Fixing the listed products script
The listed products script now works  more or less like intended, but I'm not sure how well it's gonna be on the overview page. There I'm supposed to do 10 + 10 results, not 10 -> 10 other results. But I'll get to that when I get to that

Also made CSS fixes that help my site along. Not perfect, but I'm getting where I want. More later though