Lots of CSS
I've done a few CSS changes in regards to the listed blog posts.

First off, I've made the overview page display as a grid with 1fr +1fr columns.  And I've added margins and such to make the page better on mobile.

I've also optimized the listed blog posts on the index page to look better on mobile.

Lastly I've added a loader animation to the page for when the content loads for the 1st time.