🐛 Fixed issue in the overview page
The overview page did not load more results if there were between 10 and 20 items in the array. Basically the function stopped if the count was more or less  than the length of the array.

Fixed by first of all making the count variable increase by 10 if the button is clicked (which I should have done already, but forgot to do) and secondly by running the createHTML function before disabling the "more" button and adding the "no more posts"-div.

Things are working properly now 😎