Small CSS change, QOL + loader on single posts
Small CSS change, Quality of Life; Set a min-height for the single posts on overview page. They are more uniform now.

I have also done some work in PHP to make the automatic excerpts 20 words instead of 55, but that was done in a local file, not on github.

And I added a loader on the single posts page