✨ Made the sticky post function work
Made the sticky post function work. I had to go a little caveman on it, as the "sticky=true" endpoint in the API did not link to the image.

But it had the ID to the post, which would work with the original version of createSinglePost.js.

So got the ID by looping through the array of one object which was returned in the original JSON file,  then ran the original version of  the function via that loop. Now everything works fine  😁