✨ Made prev/next post btn work | CSS faff
Made the previous and next post buttons work on single post pages.

The solution was to first get an array of all the posts, then figure out where I was in the array. Then I used array[i + 1].id; to get the ID of the next item in the list, and then I linked to the relevant URL with location.href and a dynamic value for the id

The rest of the commit is CSS faff and minor HTML changes needed to get the buttons to work. Minor changes